+++ docs/internal/processes/development-guide-relevant-files.md
@@ -162,7 +162,7 @@ This inventory records the checked-in files and directories that the maintainer 
 | `pkg/config/layout.go` | Canonical factory clone owner | Root-level `FactoryConfig` fields such as authored `factory.guards` must be copied through `CloneFactoryConfig` so `LoadRuntimeConfig` and other layout flows do not silently drop supported customer policy; hosted worker `Auth` and `Linear` deep-copy policy should stay delegated to the canonical same-package helpers in `pkg/config/agents_config.go` instead of being recopied inline here; portable expand and named-factory persistence should also keep thin imported `factory.json` generation behind `authoredFactoryConfigForExpandedLayout(...)` before flattening so regression tests assert the same authored-layout contract users actually get on disk. |
 | `pkg/config/layouttests/layout_clone_test.go` | Layout clone behavioral regression surface | `CloneWorkerConfig(...)` and `CloneFactoryConfig(...)` regression coverage should assert observable detached-copy behavior from an external test package here by mutating the source config after cloning and checking the cloned outcome, rather than asserting helper placement, helper names, or source-file inventories. |
 | `pkg/config/maptests/`, `pkg/config/mappingtests/`, `pkg/config/runtimetests/`, and `pkg/config/exhaustiontests/` | Config package-size test split seam | Keep exported-surface mapper, generated-boundary mapping, runtime-load, named-factory, and contract-guard coverage in focused external test packages when root `pkg/config` approaches the file-count cap. Keep root config tests only for package-private semantic validator, layout materialization, and staged persistence helpers that should not be exported for test movement. |
-| `pkg/service/modeltests/` | Service model API package-size test split seam | Keep black-box coverage for exported model read APIs such as `FactoryService.ListModels(...)` and `GetModel(...)` in this external test package, building the service through disk-backed factory config rather than assigning package-private runtime fields. Keep root `pkg/service` model tests only when they need package-private puller, local-runtime, resource-limiter, or invocation executor seams. |
+| `pkg/service/modeltests/` | Service model API package-size test split seam | Keep black-box coverage for exported model read APIs and direct invocation APIs such as `FactoryService.ListModels(...)`, `GetModel(...)`, and `InvokeModel(...)` in this external test package, building the service through disk-backed factory config rather than assigning package-private runtime fields. Keep root `pkg/service` model tests only when they need package-private puller, local-runtime, or resource-limiter seams. |
 | `pkg/service/factory_runtime_state.go` | Service runtime observability owner | Runtime state helpers and model-execution event recording can share this owner while the file remains below the 1000-line backend size gate; avoid reintroducing a separate root file for model event recorder helpers unless the runtime observability surface is split into a real subpackage. |
 | `pkg/service/runtimelogtests/` | Service runtime-log package-size test split seam | Keep black-box runtime-log and replay-correlation coverage that can drive `service.BuildFactoryService(...)` through exported configuration in this external test package. Keep only the small same-package test doubles and log parsers needed by remaining root service tests in `pkg/service/factory_test_helpers_test.go` until those white-box seams move behind production owners. |
 | `pkg/service/replaytests/` | Service replay package-size test split seam | Keep replay-mode service behavior that can be proven through exported `BuildFactoryService(...)`, `Run(...)`, snapshots, replay artifacts, and recorded events in this external test package. Keep shared replay event builders in the root service helper owner only while same-package record-mode or dashboard tests still consume them. |
